fixing 🐛 related to `generate`

+++ megatron/model/gpt_model.py
@@ -30,7 +30,7 @@ def post_language_model_processing(lm_output, labels, logit_weights,
         # gather along the vocab dimension
         output = mpu.gather_from_tensor_model_parallel_region(output)
         # [s b h] => [b s h]
-        return output.transpose(0,1).contiguous()
+        return output.float().transpose(0,1).contiguous()
     else:
         # [b s] => [s b]
         labels = labels.transpose(0,1).contiguous()
@@ -45,7 +45,7 @@ def post_language_model_processing(lm_output, labels, logit_weights,
         if return_logits:
             # gather along the vocab dimension
             output = mpu.gather_from_tensor_model_parallel_region(output)
-            return loss, output.transpose(0,1).contiguous()
+            return loss, output.float().transpose(0,1).contiguous()
         else:
             return loss
